categoryPanel and selectionPanel
Revamped selectionPanel in many ways, primarily in now displaying all notes of a selected tag from the category panel in the selectionPanel. Added along the way a new private class NotePreivew that extends JPanel to hold the notes that are previewed in the selectionPanel. Also added a TreeSelectionListener private class to handle tree selection listening. getNotesFromTag in the database now always return the empty string, even if no other notes correspond to a tag as well as if some do. Added a hierarchy of goals.