Created (new)Note
Created a new Note class that allows easy translation between database and app.  Adjusted much of BasicNotesDB to reflect this change.  Also added function to return notes from a category. Expanded responsibilities of simpleListener (was treeListener). Categories can now be selected and multiple categories/tags at the same time. Display of these notes has been sufficiently reworked. simpleListener now also allows the selection of a displayed note and allows editing of the related text (and displays associated tags) in the currNotePane. All of this augmented with the new Note class.